+++ com.unity.mobile.notifications/Runtime/iOS/Plugins/UnityAppController+Notifications.mm
@@ -43,30 +43,39 @@
          object: nil
          queue: [NSOperationQueue mainQueue]
          usingBlock:^(NSNotification *notification) {
-             [UnityNotificationManager sharedInstance].lastReceivedNotification = NULL;
+             UnityNotificationManager* manager = [UnityNotificationManager sharedInstance];
+             manager.launchedWithNotification = NO;
+             manager.lastReceivedNotification = NULL;
          }];
 
         [nc addObserverForName: kUnityWillFinishLaunchingWithOptions
          object: nil
          queue: [NSOperationQueue mainQueue]
          usingBlock:^(NSNotification *notification) {
-             [UNUserNotificationCenter currentNotificationCenter].delegate = [UnityNotificationManager sharedInstance];
-
-             BOOL authorizeOnLaunch = [[[NSBundle mainBundle] objectForInfoDictionaryKey: @"UnityNotificationRequestAuthorizationOnAppLaunch"] boolValue];
-             BOOL supportsPushNotification = [[[NSBundle mainBundle] objectForInfoDictionaryKey: @"UnityAddRemoteNotificationCapability"] boolValue];
+             UnityNotificationManager* manager = [UnityNotificationManager sharedInstance];
+             [UNUserNotificationCenter currentNotificationCenter].delegate = manager;
+             NSBundle* mainBundle = [NSBundle mainBundle];
+             BOOL authorizeOnLaunch = [[mainBundle objectForInfoDictionaryKey: @"UnityNotificationRequestAuthorizationOnAppLaunch"] boolValue];
+             BOOL supportsPushNotification = [[mainBundle objectForInfoDictionaryKey: @"UnityAddRemoteNotificationCapability"] boolValue];
              BOOL registerRemoteOnLaunch = supportsPushNotification == YES ?
-                 [[[NSBundle mainBundle] objectForInfoDictionaryKey: @"UnityNotificationRequestAuthorizationForRemoteNotificationsOnAppLaunch"] boolValue] : NO;
+                 [[mainBundle objectForInfoDictionaryKey: @"UnityNotificationRequestAuthorizationForRemoteNotificationsOnAppLaunch"] boolValue] : NO;
 
-             NSInteger defaultAuthorizationOptions = [[[NSBundle mainBundle] objectForInfoDictionaryKey: @"UnityNotificationDefaultAuthorizationOptions"] integerValue];
+             NSInteger defaultAuthorizationOptions = [[mainBundle objectForInfoDictionaryKey: @"UnityNotificationDefaultAuthorizationOptions"] integerValue];
 
              if (defaultAuthorizationOptions <= 0)
                  defaultAuthorizationOptions = (UNAuthorizationOptionSound + UNAuthorizationOptionAlert + UNAuthorizationOptionBadge);
 
+             if (notification != nil && notification.userInfo != nil)
+                 for (NSString* key in notification.userInfo)
+                     if ([key isEqual: @"UIApplicationLaunchOptionsLocalNotificationKey"]
+                         || [key isEqual: @"UIApplicationLaunchOptionsRemoteNotificationKey"])
+                     {
+                         manager.launchedWithNotification = YES;
+                         break;
+                     }
+
              if (authorizeOnLaunch)
-             {
-                 UnityNotificationManager* manager = [UnityNotificationManager sharedInstance];
                  [manager requestAuthorization: defaultAuthorizationOptions withRegisterRemote: registerRemoteOnLaunch forRequest: NULL];
-             }
          }];
 
         [nc addObserverForName: kUnityDidRegisterForRemoteNotificationsWithDeviceToken