+++ com.unity.mobile.notifications/Runtime/iOS/iOSNotificationsWrapper.cs
@@ -74,7 +74,7 @@ namespace Unity.Notifications.iOS
         private static extern Int32 _GetApplicationBadge();
 
         [DllImport("__Internal")]
-        private static extern bool _GetAppOpenedUsingNotification();
+        private static extern Int32 _GetAppOpenedUsingNotification();
 
         [DllImport("__Internal")]
         internal static extern void _RemoveAllDeliveredNotifications();
@@ -426,7 +426,7 @@ namespace Unity.Notifications.iOS
         public static bool GetAppOpenedUsingNotification()
         {
 #if UNITY_IOS && !UNITY_EDITOR
-            return _GetAppOpenedUsingNotification();
+            return _GetAppOpenedUsingNotification() != 0;
 #else
             return false;
 #endif
@@ -435,21 +435,18 @@ namespace Unity.Notifications.iOS
         public static iOSNotificationWithUserInfo? GetLastNotificationData()
         {
 #if UNITY_IOS && !UNITY_EDITOR
-            if (_GetAppOpenedUsingNotification())
-            {
-                IntPtr ptr = _GetLastNotificationData();
+            IntPtr ptr = _GetLastNotificationData();
 
-                if (ptr != IntPtr.Zero)
-                {
-                    iOSNotificationWithUserInfo data;
-                    data.data = (iOSNotificationData)Marshal.PtrToStructure(ptr, typeof(iOSNotificationData));
-                    data.userInfo = NSDictionaryToCs(data.data.userInfo);
-                    data.data.userInfo = IntPtr.Zero;
-                    data.attachments = AttachmentsNSArrayToCs(data.data.attachments);
-                    data.data.attachments = IntPtr.Zero;
-                    _FreeUnmanagediOSNotificationDataArray(ptr, 1);
-                    return data;
-                }
+            if (ptr != IntPtr.Zero)
+            {
+                iOSNotificationWithUserInfo data;
+                data.data = (iOSNotificationData)Marshal.PtrToStructure(ptr, typeof(iOSNotificationData));
+                data.userInfo = NSDictionaryToCs(data.data.userInfo);
+                data.data.userInfo = IntPtr.Zero;
+                data.attachments = AttachmentsNSArrayToCs(data.data.attachments);
+                data.data.attachments = IntPtr.Zero;
+                _FreeUnmanagediOSNotificationDataArray(ptr, 1);
+                return data;
             }
 #endif
             return null;