+++ com.unity.mobile.notifications/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this package will be documented in this file.
 
+## [Unreleased]
+
+- [iOS] Added a new API (QueryLastRespondedNotification) for retrieving notification details when an application is launched by tapping the notification.
+
 ## [2.3.2] - 2024-02-09
 
 ### Fixes: